[Fix]: Change type to size_t for i and j

+++ 0x1B-sorting_algorithms/0-bubble_sort.c
@@ -9,7 +9,7 @@
 
 void bubble_sort(int *array, size_t size)
 {
-	unsigned int i, j;
+	size_t i, j;
 
 	for (i = 0; i < size - 1; i++)
 	{
